@@ -8,37 +8,33 @@
   },
   "scripts": {
     "clean": "rimraf dist",
-    "install-node": "curl -sL https://deb.nodesource.com/setup_16.x | sudo -E bash",
-    "setup": "sudo npm install -g puppeteer --unsafe-perm=true -allow-root && sudo apt install chromium-browser -y && sudo npm install pm2 -g",
-    "setup-puppeteer": "sudo apt update && sudo apt install -y gconf-service libgbm-dev libasound2 libatk1.0-0 libc6 libcairo2 libcups2 libdbus-1-3 libexpat1 libfontconfig1 libgcc1 libgconf-2-4 libgdk-pixbuf2.0-0 libglib2.0-0 libgtk-3-0 libnspr4 libpango-1.0-0 libpangocairo-1.0-0 libstdc++6 libx11-6 libx11-xcb1 libxcb1 libxcomposite1 libxcursor1 libxdamage1 libxext6 libxfixes3 libxi6 libxrandr2 libxrender1 libxss1 libxtst6 ca-certificates fonts-liberation libappindicator1 libnss3 lsb-release xdg-utils wget",
-    "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "nodemon src/index.ts",
     "start": "node dist/index.js",
-    "build": "npm run clean && node build.js"
+    "build": "npm run clean && tsc"
   },
   "author": "Piyush Duggal",
   "license": "ISC",
   "dependencies": {
-    "axios": "^1.2.3",
-    "date-fns": "^2.29.3",
-    "dotenv": "^16.0.3",
+    "axios": "^1.4.0",
+    "date-fns": "^2.30.0",
+    "dotenv": "^16.1.4",
     "express": "^4.18.2",
-    "mongoose": "^6.8.4",
+    "mongoose": "^7.2.2",
     "qrcode-terminal": "^0.12.0",
     "uuid": "^9.0.0",
-    "whatsapp-web.js": "^1.19.4",
-    "winston": "^3.8.2",
+    "whatsapp-web.js": "^1.21.0",
+    "winston": "^3.9.0",
     "wwebjs-mongo": "^1.1.0"
   },
   "devDependencies": {
-    "@types/express": "^4.17.15",
+    "@types/express": "^4.17.17",
     "@types/qrcode-terminal": "^0.12.0",
-    "@types/uuid": "^9.0.0",
-    "esbuild": "^0.17.2",
-    "madge": "^5.0.1",
-    "nodemon": "^2.0.20",
+    "@types/uuid": "^9.0.2",
+    "esbuild": "^0.17.19",
+    "madge": "^6.1.0",
+    "nodemon": "^2.0.22",
     "tiny-glob": "^0.2.9",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.5"
+    "typescript": "^5.1.3"
   }
 }
