@@ -1,41 +1,41 @@
-import * as WAWebJS from 'whatsapp-web.js';
+import * as WAWebJS from "whatsapp-web.js";
 import {
   Client,
   GroupNotification,
   LocalAuth,
   MessageMedia,
-} from 'whatsapp-web.js';
-import qrcode = require('qrcode-terminal');
-import { checkMessage, superCmdFilter } from './actions/messageActions';
-import { main } from './controllers/main';
-import { introduction, sendCommands } from './actions/introduction';
+} from "whatsapp-web.js";
+import qrcode = require("qrcode-terminal");
+import { checkMessage, superCmdFilter } from "./actions/messageActions";
+import { main } from "./controllers/main";
+import { introduction, sendCommands } from "./actions/introduction";
 import {
   GREETINGS,
   HEY_EMOJIES,
   RANDOM_WAKEUP_MSG,
   USER_JOIN_GREETINGS,
-} from './utils/reply/replies';
-import { random } from './actions/sendMessage';
-import express from 'express';
-import * as dotenv from 'dotenv';
-import { Request, Response } from 'express';
-import { COMMANDS_CMDS } from './utils/Commands/instructions';
-import { grpLeaveStickers } from './assets/assets';
-import { log } from './utils/log';
-import { MessageType, WA_Grp } from './types/types';
-import { sendAndDeleteMsg } from './actions/sendAndDeleteMsg';
-import { pingEveryone } from './actions/pingEveryone';
-import { addUser, increaseNumberOfCmd, removeUser } from './services/mongo';
-import { connectToDb } from './utils/db/connect';
-import { removeMember } from './actions/removeMember';
+} from "./utils/reply/replies";
+import { random } from "./actions/sendMessage";
+import express from "express";
+import * as dotenv from "dotenv";
+import { Request, Response } from "express";
+import { COMMANDS_CMDS } from "./utils/Commands/instructions";
+import { grpLeaveStickers } from "./assets/assets";
+import { log } from "./utils/log";
+import { MessageType, WA_Grp } from "./types/types";
+import { sendAndDeleteMsg } from "./actions/sendAndDeleteMsg";
+import { pingEveryone } from "./actions/pingEveryone";
+import { addUser, increaseNumberOfCmd, removeUser } from "./services/mongo";
+import { connectToDb } from "./utils/db/connect";
+import { removeMember } from "./actions/removeMember";
 
 // @ts-ignore
 import {
   GrpJoinNotification,
   GrpLeaveNotification,
   MessageTypeOfWA,
-} from './utils/returnTypeOfWA';
-import { sendClassNotification } from './actions/sendClassNotification';
+} from "./utils/returnTypeOfWA";
+import { sendClassNotification } from "./actions/sendClassNotification";
 dotenv.config();
 
 import logger from "./utils/logger/index";
@@ -44,7 +44,7 @@ import logger from "./utils/logger/index";
 const app = express();
 
 // For Development Enviornment
-const LOCAL = String(process.env.dev) === 'true';
+const LOCAL = String(process.env.dev) === "true";
 export const BOT = LOCAL ? 1 : 0;
 export const WA_BOT_ID = LOCAL
   ? (process.env.WA_BOT_ID_DEV as string)
@@ -61,7 +61,7 @@ connectToDb(DB_URL);
 const client = new Client({
   puppeteer: {
     headless: true,
-    args: ['--no-sandbox', '--disable-setuid-sandbox'],
+    args: ["--no-sandbox", "--disable-setuid-sandbox"],
   },
   authStrategy: new LocalAuth({
     dataPath: `${__dirname}/sessions`,
@@ -69,7 +69,7 @@ const client = new Client({
 });
 
 // For QR Code
-client.on('qr', (qr: string) => {
+client.on("qr", (qr: string) => {
   qrcode.generate(qr, { small: true });
   console.log(qr);
   logger.info("QR generated", { label: "INFO" });
@@ -90,73 +90,59 @@ client.on("ready", async () => {
  * Event "MESSAGE_CREATE"
  * @returns { MessageTypeOfWA }
  */
-client.on('message_create', async (message: WAWebJS.Message) => {
+client.on("message_create", async (message: WAWebJS.Message) => {
+  console.log(message)
   // Check if message is from Group or Not, if yes, who contains whoean or userID
   const userObj: MessageType = await checkMessage(message);
   // Mention Logic
-  const str: string[] = message.mentionedIds;
-  const isMention =
-    (message.body[0] === '@' && str.includes('919871453667@c.us')) ||
-    message.body
-      .toLowerCase()
-      .split(' ')
-      .includes(`@${(process.env.BOT_NAME as String).toLocaleLowerCase()}`);
-  if (
-    isMention &&
-    userObj.role !== 'NONE' &&
-    message.body.split(' ').length === 1
-  ) {
-    introduction(client, userObj, message);
-  }
-  let allChats = await client.getChats();
-  const WA_BOT: WA_Grp = allChats[BOT];
+  // let allChats = await client.getChats();
+  // const WA_BOT: WA_Grp = allChats[BOT];
+  //
+  // const cmd = message.body.split(",")[0].toLocaleLowerCase();
 
-  const cmd = message.body.split(',')[0].toLocaleLowerCase();
-
-  // Command check logic
-  if (COMMANDS_CMDS.includes(cmd)) {
-    sendCommands(client, message, userObj);
-    await increaseNumberOfCmd({ recipitantId: userObj.chatId });
-    return;
-  }
-  if (userObj.role === 'STUDENT' && superCmdFilter(message.body)) {
-    WA_BOT.sendMessage(
-      'You cannot perform this action, because you are not a BOT ADMIN, you will get ban if you use this frequently :)'
-    );
-    return;
-  }
-
-  if (userObj.role !== 'NONE' && superCmdFilter(message.body)) {
-    console.log('entering removing');
-    await removeMember(WA_BOT as WAWebJS.GroupChat, userObj, message);
-    return;
-  }
-
-  // Ping Everyone
-  if (userObj.role === 'OWNER' && ['everyone'].includes(message.body)) {
-    await pingEveryone(client, message);
-    return;
-  }
+  // // Command check logic
+  // if (COMMANDS_CMDS.includes(cmd)) {
+  //   sendCommands(client, message, userObj);
+  //   await increaseNumberOfCmd({ recipitantId: userObj.chatId });
+  //   return;
+  // }
+  // if (userObj.role === "STUDENT" && superCmdFilter(message.body)) {
+  //   WA_BOT.sendMessage(
+  //     "You cannot perform this action, because you are not a BOT ADMIN, you will get ban if you use this frequently :)"
+  //   );
+  //   return;
+  // }
+  //
+  // if (userObj.role !== "NONE" && superCmdFilter(message.body)) {
+  //   console.log("entering removing");
+  //   await removeMember(WA_BOT as WAWebJS.GroupChat, userObj, message);
+  //   return;
+  // }
+  //
+  // // Ping Everyone
+  // if (userObj.role === "OWNER" && ["everyone"].includes(message.body)) {
+  //   await pingEveryone(client, message);
+  //   return;
+  // }
 
   // Checks if message's first letter is BOT_PREFIX
-  if (
-    userObj.role !== 'NONE' &&
-    message.body[0] === (process.env.BOT_PREFIX as string)
-  ) {
+  // if (
+  //   userObj.role !== "NONE"
+  // ) {
     await main(client, message);
     return;
-  }
+  // }
   // WARN: ONLY USE ONCE
-  if (userObj.role === 'OWNER' && message.body === 'load') {
-    WA_BOT.participants?.forEach(async (participant) => {
-      let recipitantId = participant.id._serialized;
-      await addUser({ recipitantId });
-    });
-    allChats[1].sendMessage(
-      'SUCCESSFULLY ADDED ALL THE STUDENTS IN THE DB, MASTER!'
-    );
-    return;
-  }
+  // if (userObj.role === "OWNER" && message.body === "load") {
+  //   WA_BOT.participants?.forEach(async (participant) => {
+  //     let recipitantId = participant.id._serialized;
+  //     await addUser({ recipitantId });
+  //   });
+  //   allChats[1].sendMessage(
+  //     "SUCCESSFULLY ADDED ALL THE STUDENTS IN THE DB, MASTER!"
+  //   );
+  //   return;
+  // }
 });
 
 /**
@@ -164,15 +150,15 @@ client.on('message_create', async (message: WAWebJS.Message) => {
  * Event "GROUP_JOIN"
  * @returns { GrpJoinNotification }
  */
-client.on('group_join', async (msg: GroupNotification) => {
+client.on("group_join", async (msg: GroupNotification) => {
   if (msg.chatId === WA_BOT_ID) {
     log({
       msg: `${msg.recipientIds[0]} Joined the Group`,
-      type: 'GROUP_JOIN',
+      type: "GROUP_JOIN",
       error: false,
     });
     const contact = await client.getNumberId(msg.recipientIds[0]);
-    const details = await client.getContactById(contact?._serialized || '');
+    const details = await client.getContactById(contact?._serialized || "");
     if (details.name) {
       sendAndDeleteMsg(
         client,
@@ -239,16 +225,16 @@ client.on('group_join', async (msg: GroupNotification) => {
  * Event "GROUP_LEAVE"
  * @returns { GrpLeaveNotification }
  */
-client.on('group_leave', async (notification: WAWebJS.GroupNotification) => {
+client.on("group_leave", async (notification: WAWebJS.GroupNotification) => {
   let grpId = notification.chatId;
   if (grpId === WA_BOT_ID) {
     log({
       msg: `${notification.recipientIds[0]} left the Group`,
-      type: 'GROUP_LEFT',
+      type: "GROUP_LEFT",
       error: false,
     });
   }
-  if (notification.chatId === WA_BOT_ID && notification.type !== 'remove') {
+  if (notification.chatId === WA_BOT_ID && notification.type !== "remove") {
     const sticker = MessageMedia.fromFilePath(
       `${__dirname}/../src/assets/images/grpJoinLeaveImgs/${
         grpLeaveStickers.images[random(grpLeaveStickers.numOfImgs)]
@@ -286,17 +272,17 @@ client.initialize();
 //   clearInterval(intervalId);
 // }, etaMs);
 
-const port = Number(process.env.PORT) || 3005;
+const port = Number(process.env.PORT) || 3000;
 
-app.get('/', (_: Request, res: Response) => {
-  res.send('BOT');
+app.get("/", (_: Request, res: Response) => {
+  res.send("BOT");
 });
 app.listen(port, () =>
   logger.info(`[SERVER] Server is running on port ${port}`, { label: "INFO" })
 );
 
 // All other pages should be returned as error pages
-app.all('*', (_: Request, res: Response) => {
+app.all("*", (_: Request, res: Response) => {
   res
     .status(404)
     .send(
